#515 refactor on sport club table

+++ aaran/SportsClub/Database/Migrations/2024_03_01_004009_create_sport_highlights_table.php
@@ -4,24 +4,25 @@ use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
-return new class extends Migration
-{
+return new class extends Migration {
 
     public function up(): void
     {
-        Schema::create('sport_highlights', function (Blueprint $table) {
-            $table->id();
-            $table->foreignId('activity_id')->references('id')->on('sport_activities')->onDelete('cascade');
-            $table->string('vname');
-            $table->longText('url')->nullable();
-            $table->string('active_id', 3)->nullable();
-            $table->foreignId('sport_club_id')->references('id')->on('sport_clubs');
-            $table->foreignId('tenant_id')->references('id')->on('tenants');
-            $table->foreignId('user_id')->references('id')->on('users');
-            $table->timestamps();
-        });
-    }
+        if (Aaran\Aadmin\Src\DbMigration::hasSports()) {
 
+            Schema::create('sport_highlights', function (Blueprint $table) {
+                $table->id();
+                $table->foreignId('activity_id')->references('id')->on('sport_activities')->onDelete('cascade');
+                $table->string('vname');
+                $table->longText('url')->nullable();
+                $table->string('active_id', 3)->nullable();
+                $table->foreignId('sport_club_id')->references('id')->on('sport_clubs');
+                $table->foreignId('tenant_id')->references('id')->on('tenants');
+                $table->foreignId('user_id')->references('id')->on('users');
+                $table->timestamps();
+            });
+        }
+    }
 
     public function down(): void
     {
